finish testing add and remove todo
