+++ package.json
@@ -4,17 +4,16 @@
   "description": "Javascript to do list.",
   "private": "true",
   "scripts": {
-    "test": "jest --coverage --verbose true",
+    "test": "jest --env=jsdom",
     "start": "webpack serve --open",
     "server": "node server.js",
     "build": "webpack"
   },
   "jest": {
-    "testEnvironment": "jsdom",
-    "moduleNameMapper":{
+    "moduleNameMapper": {
       "\\.(css|less|sass|scss)$": "<rootDir>/__mocks__/styleMock.js",
       "\\.(gif|ttf|eot|svg)$": "<rootDir>/__mocks__/fileMock.js"
- }
+    }
   },
   "repository": {
     "type": "git",
@@ -28,8 +27,11 @@
   },
   "homepage": "https://github.com/bill7pearl/todo#readme",
   "devDependencies": {
+    "@babel/core": "^7.20.5",
     "@babel/plugin-transform-modules-commonjs": "^7.19.6",
+    "@babel/preset-env": "^7.20.2",
     "babel-eslint": "^10.1.0",
+    "babel-loader": "^9.1.0",
     "css-loader": "^6.7.2",
     "eslint": "^7.32.0",
     "eslint-config-airbnb-base": "^14.2.1",
@@ -40,6 +42,7 @@
     "html-webpack-plugin": "^5.5.0",
     "image-webpack-loader": "^4.2.0",
     "jest": "^29.3.1",
+    "jest-environment-jsdom": "^29.3.1",
     "mini-css-extract-plugin": "^2.7.2",
     "style-loader": "^3.3.1",
     "stylelint": "^13.13.1",
@@ -53,7 +56,5 @@
   },
   "main": "webpack.config.js",
   "dependencies": {
-    "jsdom": "^20.0.3"
   }
-  
 }
