+++ src/index.html
@@ -17,7 +17,7 @@
 					placeholder="Please type your todo here..."
 					name="content"
 					id="content" />
-				<input type="submit" value="Add todo" />
+				<input class="addInput" type="submit" value="Add todo" />
 			</form>
 		</section>
