--- src/modules/sum.js
@@ -1,4 +1,0 @@
-function sum(a, b) {
-  return a + b;
-}
-module.exports = sum;
