--- src/modules/sum.test.js
@@ -1,5 +1,0 @@
-const sum = require('./sum');
-
-test('adds 1 + 2 to equal 3', () => {
-  expect(sum(1, 2)).toBe(3);
-});
